Added match check to term.read
It can be ethier string or function. If it's string user can confirm only when entered text matches the string. If thes param is function, the function gets called when user clicks enter key, first parameter for it is entered text, it should return a value that decides if given text should is accepted(true), or not(false/nil)