Fix InventoryUtils.insertIntoInventory
According to the documentation inventory_controller.suckFromSlot, robot.suck and
friends try to put the sucked item into the selected slot, and only if it's
occupied try others. However the implementation first tried to merge the stack
with existing stacks in the inventory, and only if it failed put into the
selected slot. This commit fixes it.