Actually string parameter for keyboard.isKeyDown()
keyboard.isKeyDown() accepts a string, but doesn't do anything sensible in the case a string is passed.

keyboard.pressedChars[] is always indexed by the unicode codepoint (maps to ASCII in the simple case), so passing a string currently always returns nil.

The proposed tweak changes that. Alternatively, we shouldn't be allowing strings in the first place, I guess.