Validate NBT address of FileSystems
Validate the address stored in hard disks and floppy disks and generate a new address if it is now longer a UUID.
Fixes breaking out of the sandbox and accessing host files with potential writes.

This does not fix being able to modify drives using a debug card or creative mode to access other drives.