Parse Lua REPL inputs with an implicit "return "
If an input does not start with a leading "=", this will parse the input
with "return " appended and, if that fails, will parse as a normal
statement.

This allows for normal expressions to be entered into the repl (such as
`2 + 2`) but does mean the parse errors for malformed inputs are
confusing. For instance, `3 + ` will error at '3' rather than '<eof>'.