Fix runClient and runServer Gradle tasks
Compile dependencies are placed on the runtime classpath by Gradle.
Because the compile configuration is set to extend from the provided
config, all of those dependencies are put on the runtime classpath as
well.

In this patch, provided is instead made a separate configuration, and
explicitly added to ONLY the compile classpath of the main sourceset.