Fix wrong Gradle version being used on Travis CI
Using the default Gradle install on Travis is dangerous, because a
version upgrade on their end could break the build. It's already broken
because the version of ForgeGradle the current script relies on expects
ScalaCompileOptions.setUseAnt to exist, but it was removed in Gradle 3.

This patch reconfigures Travis to use the Gradle wrapper instead, which
standardizes the version in use.