Added Robot names (#2555)
Got a little creative with some of them.. I have been interested in robots and robot movies for a while and got kind of caught in it. Most of robots/AIs I thought of were already in the list, but I think these would make some nice additions to the list.