Fixed drones not reclaiming their chunk tickets
If a ticket is created with ForgeChunkManager.Type.ENTITY and bound to
an entity, then forge will actually load that entity's chunk before
returning its ticket. As a result, the drones were initializing their
components before the ticket was added to restoredTickets, so they
never actually reclaimed their ticket. Also added a log message to
detect if something like this happens again.