allow redstone setOutput to specify large values
This appears to work fine with other mods, and with vanilla redstone
tested from 1.7.10 and up.

Also, I reviewed code in other mods setting redstone signals and it
seems that we don't need to cap our redstone signal

Note that this is not an exploit on vanilla redstone, setting output to
2,000 (for example) does not propogate a powered redstone "wire" (dust
line) further than the vanilla 15 blocks. Also, when reading a vanilla
redstone dust line that was powered with a value greater than 15 (e.g.
2,000 again), we do not read huge values either. The vanilla mc system
is capping the values for us.

closes #2058