allow host to build component list even if off
fixes confusing analyzer report. when you place a case in the world and
add components, even while the case is off it correctly reports its
internal component max. But when loading a world, cases that are not
running will have a 0 component max according to the analyzer. This
inaccurate limit is fixed when the case is turned on