fix more inventory interactions
1. the agent armor inventory maps to the player equipment inventory.
previously we have been preparing only the first(0) equipment(belt) item
out of robots/drones for inventory interactions. but the player
inventory has 4 slots for the armor set, enough room to represent all
of the belt from the agent. This change prepares all 4 armor positions.

2. when our player inventory is modified via the our container slots the
agent inventory is updated. this causes underlying player inventory
lists to thus become outdated. We can detect these changes by hooking
into a listener that will notify us when slots change the agents, thus
allowing us to keep the player inventory lists updated. When an
operation on the inventory container is done, we are then up to date
between the agent and the player

3. Previously we had been holding the active slot index contents in a
private offHand tuple, and previously when the item action had completed
we would detect changes to the offHand data. This extra field wasn't
needed, this change just uses the offhand inventory.

4. cleaned up wasteful code in InventoryUtils.insertIntoInventory

todo: the inventory container is a PlayerContainer, which builds various
slots we never need, and not enough slots. This causes the slot change
detection code to only cover a tiny subset of agent slots

note: the offhand solution isn't infallable. what if an item is
naturally supposed to create duplicate of itself on use. the dup would
stack with the originating slot, and the offhand data would overwrite it
when the item use action completed. but this is unlikely. a solution
would require a truly blocking item be put in place of the originating
slot as a placeholder for when the offhand action completed.