new filesystem device feature, locked mode
A locked filesystem is readonly (in either managed or unmanaged modes)
It cannot be unlocked unless its mode is switched or it is recrafted -
both actions wipe the drive. The data is thus "locked" or protected
unless the drive is wiped. The player's displace name is also recorded
and shown in tooltips on the device to indicate who locked it. In this
manner, data authenticity can be trusted

closes #2138