Fix server-side deadlock in filesystem activity handling
We've managed to hit a reproducible deadlock concerning filesystems. The exact setup is unknown, but using finer lock scopes should fix the issue. I'm not able to test this change at the moment (I don't have MCP setup), and I can't verify if the change is valid Scala (but it seems like it should be fine).