Simulating the insertion instead of inserting directly
Previously the extraction was simulated and based on the simulated
extraction the extracted items were inserted. This was problematic, as
it relyed on the fact, that the simulation returns the same result as
the actual extraction. This caused problems, as extracting a component
only saves the changes to the component, when it is not simulated.
Furthermore it could cause item deduplication, if the actual extraction
returned something different.

To solve these issues, the insertion is also simulated on the simulated
extraction and the actual insertion is done on the actual extracted
items.